refactor: 변수명 수정 및 디렉토리 변경
