+++ cookie/src/main/java/com/cookie/global/oauth/CustomSuccessHandler.java
@@ -47,14 +47,14 @@ public class CustomSuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         if (isRegistrationRequired) {
             log.warn("사용자 등록이 필요함");
             String redirectUrl = clientUrl + "/register"
-                    + "?provider=" + customUserDetails.getSocialProvider()
+                    + "?socialProvider=" + customUserDetails.getSocialProvider()
                     + "&email=" + customUserDetails.getEmail()
                     + "&socialId=" + customUserDetails.getSocialId();
 
             response.sendRedirect(redirectUrl);
         } else {
             log.info("기존 사용자 로그인");
-            String token = jwtUtil.createJwt(nickname, role, 60 * 60 * 60L);
+            String token = jwtUtil.createJwt(nickname, role);
             response.addCookie(createCookie("Authorization", token));
 
             response.sendRedirect(clientUrl + "/retrieve-token");
